add left sliding
