--- src/Components/TopNav/ShowCollection/ShowCollection.scss
@@ -1,167 +1,0 @@
-.ShowCollection {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  .collectionTxtNSwipe {
-    width: 100%;
-    .collectionSwipe {
-      height: 380px;
-      .collections {
-        position: relative;
-        width: 100%;
-        height: 100%;
-        display: flex;
-        -webkit-box-align: center;
-        align-items: center;
-        -webkit-box-pack: center;
-        .ScrollbarsCustom {
-          width: 100%;
-          height: 100%;
-
-          & > .ScrollbarsCustom-Wrapper {
-            top: 0;
-            right: 0;
-            bottom: 0;
-            left: 0;
-
-            & > .ScrollbarsCustom-Scroller {
-              & > .ScrollbarsCustom-Content {
-                //   width: 100%;
-                text-align: initial;
-                white-space: nowrap;
-                position: relative;
-                a {
-                  color: inherit;
-                  display: inline-block;
-                  width: 13vw;
-                  margin-left: 10px;
-                  &:first-child {
-                    margin-left: 8vw;
-                  }
-                }
-                li {
-                  display: inline-block;
-                  // height: 500px;
-                  width: 100%;
-
-                  overflow: hidden;
-                  &:last-child {
-                    margin-right: 8vw;
-                  }
-                  .collectionPic {
-                    height: 29.5vh;
-                    //background-image: url(https://content.rolex.com/dam/homepage/roller/all-watches/watches_yacht-master_m116680_0002_1809jva_001.jpg?imwidth=550);
-                    object-fit: cover;
-                    background-size: cover;
-                    transform: scale(1);
-                    transition: transform 0.6s ease 0s;
-                  }
-                  .collectionPicHover {
-                    height: 30vh;
-                    //background-image: url(https://content.rolex.com/dam/homepage/roller/all-watches/watches_yacht-master_m116680_0002_1809jva_001.jpg?imwidth=550);
-                    object-fit: cover;
-                    background-size: cover;
-                    transform: scale(1.05);
-                    transition: transform 0.6s ease 0s;
-                  }
-                  .collectionTitle {
-                    display: flex;
-                    flex-direction: column;
-                    height: 66.2px;
-                    transform: translateY(0px);
-                    transition: transform 0.2s ease 0s;
-                    span {
-                      &:first-child {
-                        font-weight: 700;
-                        font-size: calc(
-                          0.625rem + 0.125 * ((100vw - 20rem) / 100)
-                        );
-                        margin: 10px 0px 0px;
-                      }
-                      &:nth-child(2) {
-                        letter-spacing: 0.3px;
-                        font-size: 0.75rem;
-                        line-height: 1.3;
-                        margin: 0px;
-                      }
-                      &:last-child {
-                        opacity: 0;
-                        transition: opacity 0.2s linear 0s;
-                        padding-left: 0px;
-                        font-weight: 400;
-                        color: rgb(33, 33, 33);
-                        margin-top: 5px;
-                        font-size: 0.875rem;
-                        letter-spacing: 0.3px;
-                      }
-                    }
-                  }
-                  .collectionTitleHover {
-                    display: flex;
-                    flex-direction: column;
-                    transform: translateY(-17px);
-                    transform-origin: 50% 50%;
-                    transition: transform 0.2s ease 0s;
-                    span {
-                      &:first-child {
-                        font-weight: 700;
-                        font-size: calc(
-                          0.625rem + 0.125 * ((100vw - 20rem) / 100)
-                        );
-                        line-height: 1.4;
-                        padding: 10px 0px 0px;
-                        background-color: rgb(33, 33, 33);
-                      }
-                      &:nth-child(2) {
-                        letter-spacing: 0.3px;
-                        font-size: 0.75rem;
-                        line-height: 1.3;
-                        margin: 0px;
-                      }
-                      &:last-child {
-                        opacity: 1;
-                        transition: opacity 0.2s linear 0s;
-                        padding-left: 0px;
-                        font-weight: 400;
-                        margin-top: 4.5px;
-                        font-size: 0.875rem;
-                        letter-spacing: 0.3px;
-                      }
-                    }
-                  }
-                }
-                &::-webkit-scrollbar {
-                  display: none;
-                }
-              }
-            }
-          }
-
-          .ScrollbarsCustom-Track {
-            border-radius: 4px;
-          }
-          .ScrollbarsCustom-ThumbX {
-            height: 4px;
-            border-radius: 4px;
-            background: #ffffff;
-            transition: height 0.2s ease 0s;
-            &:hover {
-              height: 5px;
-              transition: height 0.2s ease 0s;
-            }
-          }
-          .ScrollbarsCustom-TrackX {
-            position: absolute;
-            display: flex;
-            align-items: center;
-            height: 4.5px;
-            width: calc(100% - 16vw);
-            bottom: 0;
-            left: 8vw;
-            background-color: rgba(255, 255, 255, 0.2);
-          }
-        }
-      }
-    }
-  }
-}
