+++ src/Components/TopNav/inMenu.scss
@@ -36,6 +36,7 @@
       position: absolute;
       margin-left: 80px;
       margin-top: 50px;
+      z-index: 7;
       .cancel {
         fill: white;
         &:hover {
@@ -50,14 +51,21 @@
       height: 100%;
       background-color: #212121;
       overflow: scroll;
-      padding-left: 130px;
-      padding-right: 130px;
+      padding-top: 100px;
       .first {
-        width: 40%;
-        height: 80px;
+        position: absolute;
+        top: -100px;
+        width: 50%;
+        height: 30%;
         z-index: 5;
         background-color: #212121;
-        position: fixed;
+      }
+      .slideShow {
+        margin-left: -30px;
+        padding-top: 20px;
+        width: 100%;
+        height: 50%;
+        z-index: 50;
       }
       .st0 {
         fill: #a37e2d;
@@ -75,11 +83,8 @@
         cursor: pointer;
       }
     }
-    .slideShow {
-      padding-top: 200px;
-    }
     .menuList {
-      padding: 30px 0;
+      padding: 30px 80px;
       list-style-type: none;
       text-align: left;
       font-size: 1rem;
@@ -122,3 +127,171 @@
   background-color: black;
   opacity: 0.9;
 }
+
+// .collections {
+//   position: relative;
+//   width: 100%;
+//   height: 100%;
+//   //height: 473.763px;
+//   display: flex;
+//   -webkit-box-align: center;
+//   align-items: center;
+//   -webkit-box-pack: center;
+
+//   .ScrollbarsCustom {
+//     width: 100%;
+//     height: 100%;
+//     .ScrollbarsCustom-Wrapper {
+//       top: 0;
+//       right: 0;
+//       bottom: 0;
+//       left: 0;
+//       .ScrollbarsCustom-Scroller {
+//         scroll-behavior: smooth;
+//         .ScrollbarsCustom-Content {
+//           width: 100%;
+//           text-align: initial;
+//           white-space: nowrap;
+//           position: relative;
+//           a {
+//             color: inherit;
+//             display: inline-block;
+//             width: 377px;
+//             margin-left: 22px;
+//             &:first-child {
+//               margin-left: 5vw;
+//             }
+//             &:last-child {
+//               margin-right: 5vw;
+//             }
+//           }
+//           li {
+//             display: inline-block;
+//             //position: relative;
+//             height: 490.79px;
+//             width: 100%;
+
+//             overflow: hidden;
+
+//             .collectionPic {
+//               height: 424.59px;
+//               //background-image: url(https://content.rolex.com/dam/homepage/roller/all-watches/watches_yacht-master_m116680_0002_1809jva_001.jpg?imwidth=550);
+//               object-fit: cover;
+//               background-size: cover;
+//               transform: scale(1);
+//               transition: transform 0.6s ease 0s;
+//             }
+//             .collectionPicHover {
+//               height: 424.59px;
+//               //background-image: url(https://content.rolex.com/dam/homepage/roller/all-watches/watches_yacht-master_m116680_0002_1809jva_001.jpg?imwidth=550);
+//               object-fit: cover;
+//               background-size: cover;
+//               transform: scale(1.05);
+//               transition: transform 0.6s ease 0s;
+//             }
+//             .collectionTitle {
+//               display: flex;
+//               flex-direction: column;
+//               height: 66.2px;
+//               background-color: white;
+//               transform: translateY(0px);
+//               transition: transform 0.2s ease 0s;
+//               span {
+//                 &:first-child {
+//                   font-weight: 700;
+//                   font-size: calc(0.6875rem + 0.3125 * ((100vw - 20rem) / 100));
+//                   line-height: 1.4;
+//                   margin: 10px 0px 0px;
+//                 }
+//                 &:nth-child(2) {
+//                   letter-spacing: 0.3px;
+//                   font-size: 0.875rem;
+//                   line-height: 1.3;
+//                   margin: 0px;
+//                 }
+//                 &:last-child {
+//                   opacity: 0;
+//                   transition: opacity 0.2s linear 0s;
+//                   padding-left: 0px;
+//                   font-weight: 400;
+//                   color: rgb(33, 33, 33);
+//                   margin-top: 5px;
+//                   font-size: 0.875rem;
+//                   letter-spacing: 0.3px;
+//                 }
+//               }
+//             }
+//             .collectionTitleHover {
+//               display: flex;
+//               flex-direction: column;
+//               //height: 66.2px;
+//               background-color: white;
+//               transform: translateY(-10px);
+//               transform-origin: 50% 50%;
+//               transition: transform 0.2s ease 0s;
+//               //overflow: hidden;
+//               span {
+//                 &:first-child {
+//                   font-weight: 700;
+//                   font-size: calc(0.6875rem + 0.3125 * ((100vw - 20rem) / 100));
+//                   line-height: 1.4;
+//                   margin: 10px 0px 0px;
+//                 }
+//                 &:nth-child(2) {
+//                   letter-spacing: 0.3px;
+//                   font-size: 0.875rem;
+//                   line-height: 1.3;
+//                   margin: 0px;
+//                 }
+//                 &:last-child {
+//                   opacity: 1;
+//                   transition: opacity 0.2s linear 0s;
+//                   padding-left: 0px;
+//                   font-weight: 400;
+//                   color: rgb(33, 33, 33);
+//                   margin-top: 5px;
+//                   font-size: 0.875rem;
+//                   letter-spacing: 0.3px;
+//                 }
+//               }
+//             }
+//           }
+//           &::-webkit-scrollbar {
+//             display: none;
+//           }
+//         }
+//       }
+//     }
+
+//     .ScrollbarsCustom-Track {
+//       border-radius: 4px;
+//     }
+
+//     .ScrollbarsCustom-Thumb {
+//       background: rgba(255, 255, 255, 0.12);
+//       cursor: pointer;
+//       border-radius: 4px;
+//     }
+
+//     .ScrollbarsCustom-ThumbX {
+//       height: 4px;
+//       background: rgb(18, 120, 74);
+//       transition: height 0.2s ease 0s;
+//       &:hover {
+//         height: 6px;
+//         transition: height 0.2s ease 0s;
+//       }
+//     }
+
+//     .ScrollbarsCustom-TrackX {
+//       position: absolute;
+//       display: flex;
+//       align-items: center;
+//       height: 4px;
+//       width: calc(100% - 10vw);
+//       bottom: 0;
+//       left: 5vw;
+//       background-color: rgba(33, 33, 33, 0.2);
+//     }
+//   }
+// }
